3.6.5 Release Date: 2019-03-01 New Final Report
3.6.6
Release Date: 2019-04-23
Allow sorting of Test Request Forms (TRFs) by the Description field on the Data Entry TRF Attachment screen.
Allow adding Patient Notes to patients linked to released samples without needing to un-release the samples.
Added the PWS Information option to the Send-out Order screen.
Allow a sample’s Collected Time to be entered without a Collected Date during full sample login.
Added Active, Begin Date, and End Date columns to the QA List Report.
Ignore QA result-overridden 2nd Trimester analyses in the MS Integrated Match Report.
Added red Content Indicators to the specified tabs on the following screens:
    Send-out Order: Organization, Aux Data, Test, Container, Item, Shipping, Customer, Internal, Sample, Filled, Attachment
    Organization: Contact, Parameter, Note, Attachment
    Worksheet Builder: Note
    Worksheet Completion: Note, Attachment
    QC: Internal, Attachment
    Instrument: Attachment
Added the capability to add a test by entering its Test Id and display the Test Id on additional screens.
Increased the maximum number of characters that can be entered into a worksheet result field.
Added Active checkbox, Begin Date, and End Date fields to the QA Event screen.
Added the capability to duplicate an auxiliary (aux) group through the Auxiliary Prompt screen.
Added “SH” to the barcode on Shipping Manifests.
Clear the releaser's username during analysis unrelease.
Display the Patient Middle Name as part of the Clinical label on the Sample Tracking tree.
Display the Project dropdown list on the OpenELIS Web Portal query screens by ascending project name instead of ascending project id.
Display the Instrument Name dropdown list on the Worksheet Builder and Worksheet Completion screens by ascending instrument name instead of ascending instrument id.
Moved the shipping method below the cost center on the shipping label so that it doesn't overlay on a long organization name.
Corrected the display of the fields at the bottom of the Items tab on the Inventory Receipt screen so that they aren't partially hidden.
Display only auxiliary (aux) data linked to samples and not to the send-out orders when using the “Result: Exclude All” filter on the Data View screen.
Updated the Chain of Custody/Tracking Signatures on the State Hygienic Laboratory’s Environmental Sample Collection Form.
Changes were made to the validation performed by the Maternal Screening test scriptlets.
Added a test scriptlet for misc radchem, miscellaneous (Radiochemical Analysis, Miscellaneous) test.
These changes were made to the Missouri State Public Health Laboratory’s CT-GC test scriptlet:
Added a scriptlet to automatically add the "coll dt blank - int" QA event to clinical samples that are missing a Collected Date.
If the scriptlet needs to add the QA event and can't find it because there's no active QA event in the system with the name "coll dt blank - int" on trying to commit the sample, the scriptlet will show an error which tells the user that that specific QA event couldn't be found and not allow committal. For the sample to be committed, the user will either have to enter the collection date or add or activate a QA event with that name.